fix: move i18n file to root directory
