+++ i18n.ts
@@ -1,8 +1,8 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 
-import translationEN from '../locales/en/translation.json';
-import translationZH from '../locales/zh/translation.json';
+import translationEN from './locales/en/translation.json';
+import translationZH from './locales/zh/translation.json';
 
 const resources = {
     en: {
